+++ rustdoc-json/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 * Add `Builder::package_target` and `PackageTarget`
+* Correctly determine json path for `Builder::default().package("crate@1.0.0")`
 
 ## v0.7.3
 * Derive `Clone` for `rustdoc_json::Builder`